requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==5.3.1
numpy==2.1.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
realtime==2.6.0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
